backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlmodel
pydantic
SQLAlchemy
requests
chromadb
scikit-learn
numpy
librosa
soundfile
nltk
python-dotenv
openai-whisper
pyannote.audio
torch
transformers
pydub
ffmpeg-python
moviepy
